requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.4
django-haystack==3.2.1
elasticsearch==7.17.3
sqlparse==0.4.2
urllib3==1.26.9
